We need to know if executeBadge actually works or not
Our app needs to know if the library was able to put the
unread badge or not, applyCount should return true only
when it was able to put the badge and this is not the case now.

Right now most badgers sendBroadcast which will never throw any
exception if there are no receivers to handle the broadcast.
So I wrapped all sendBroadcast calls with a check and throw an
exception if there is no receiver that can handle this intent.

NovaHomeBadger was alsow hiding the failure if TeslaUnread is
not installed which is not correct.